requirements.txt
flask
flask-cors
mysql-connector-python
openai
python-dotenv
